requirements.txt
discord-py-interactions
dotenv
requests
aiosqlite
flask